requirements.txt
fastapi[all] == 0.77.1
inject == 4.3.1
click == 8.1.3
dacite == 1.6.0
pyahocorasick == 1.4.4
pandas
transformers>=4.6.0,<5.0.0
sentence-transformers
uvicorn
requests
python-dotenv
nebula2-python
aiohttp